feat: correção da requisição da latitude e longitude e outros ajustes

+++ hibisco-ws/src/main/java/enterprise/hibisco/hibiscows/rest/positionstack/PositionStackResponse.java
@@ -2,9 +2,12 @@ package enterprise.hibisco.hibiscows.rest.positionstack;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 public class PositionStackResponse {
     @Getter
-    private Double latitude;
+    private Float latitude;
+
     @Getter
-    private Double longitude;
+    private Float longitude;
 }
